local

+++ src/trainNaturalImages.py
@@ -9,7 +9,6 @@ import argparse
 import os
 from tqdm import tqdm
 import random
-from efficientnet_pytorch.gen_odin_model import CosineSimilarity
 import ipdb
 
 abs_path = '/home/ferles/medusa/src/'
@@ -33,8 +32,6 @@ def train(args):
         flag = True
 
     model = build_model(args)
-    # if 'gen' in training_configurations.checkpoint:
-    #     model._fc_nominator = CosineSimilarity(feat_dim=1280, num_centers=training_configurations.out_classes)
     model = model.to(device)
 
     dataset = args.dataset.lower()
@@ -125,8 +122,10 @@ def train(args):
 
         if epoch_val_accuracy > best_val_acc:
             best_val_acc = epoch_val_accuracy
-            torch.save(model.state_dict(), f'/raid/ferles/checkpoints/eb0/{dataset}/{training_configurations.checkpoint}.pth')
-
+            if os.path.exists('/raid/ferles/'):
+                torch.save(model.state_dict(), f'/raid/ferles/checkpoints/eb0/{dataset}/{training_configurations.checkpoint}.pth')
+            else:
+                torch.save(model.state_dict(), f'/home/ferles/checkpoints/eb0/{dataset}/{training_configurations.checkpoint}.pth')
             # if best_val_acc - checkpoint_val_accuracy > 0.05:
             #     checkpoint_val_accuracy = best_val_acc
             #     torch.save(model.state_dict(), f'/raid/ferles/checkpoints/eb0/{dataset}/{training_configurations.checkpoint}_epoch_{epoch}_accuracy_{best_val_acc}.pth')
